Goods, orders and trades: the kernel of S1.05
Goods are real-asset instruments per (product, grade class, zone), issued the
first time something names them, with each deposit's right to extract in the
same ledger table (GDS.1-3). Handlers read their own goods, rights, deliveries
and the marks at their place (architecture 4.9) and ask for transformations and
orders by intents: transformations apply at their apply point, a deposit's
depletion written by GEO alone (GEO.12); orders are admitted at 5d in lots of
the product's least quantity times the row's twins, so trade money is whole and
each twin's share stays whole (REP.9). Market kinds are declared by systems and
instantiated per good; calls cut fills to whole lots, posted meetings between
firms match in lots of both parties; trades settle all or none in stage 7 after
the dues under the kernel's `traded` reason, whose declared cost effect makes
the seller's cost of goods an expense (FRM.13). MarketId widens to u32 for the
instances. Nothing trades yet: firms hold no goods until the Stage 1 opening.

+++ crates/foundation/phx-id/src/ids.rs
@@ -40,7 +40,7 @@ id!(
     /// An instrument keeps its identity after it is retired, so the identity is never handed out twice.
     InstrumentId(u32)
 );
-id!(MarketId(u16));
+id!(MarketId(u32));
 id!(TileId(u32));
 id!(ZoneId(u32));
 id!(RegionId(u16));
